Final Push - Ready for Review
(realized never correctly renamed HTML file). Form complete. The fields now all have their labels, which are linked correctly, Added data and select lists whenever needed. Added validation in various ways throughout. Added CSS to change background color of the focused field, and whenever the field is either valid or invalid, the field changes color to either a red or green, and has a image of a checkmark or red X depending on validation. Noticed the address in the books picture is a textarea, which does not work with the given CSS styles, so I assumed it was an old picture.